first ish fully working version. hooked to front to back

+++ backend/src/http/middlewares/content_validation/schemas.ts
@@ -13,5 +13,6 @@ export type TaskId = z.infer<typeof task_id>;
 export const board_id = z.object({ board_id: z.uuid() });
 export type BoardId = z.infer<typeof board_id>;
 
-export const position = z.object({ position: z.number().positive() });
+// Positions are 1-based, matching how they are stored in the DB.
+export const position = z.object({ position: z.number().int().positive() });
 export type Position = z.infer<typeof position>;
